Add direct IO stress tests
This creates a temporary ext4 image to use for the direct IO stress tests.
The testcases in LTP blindly assume /tmp is a real filesystem and not tmpfs,
which doesn't support direct IO.  So, we use this loopback filesystem and
it exercises both the loopback and ext4 code paths.  Eventually we can add
other filesystems or something